+++ assets/javascripts/kitten/index.js
@@ -12,6 +12,7 @@ export { ButtonIcon } from './components/buttons/button-icon'
 export {
   ButtonImageWithTextAndBadge,
 } from './components/buttons/button-image-with-text-and-badge'
+export { ButtonGroup } from './components/buttons/button-group'
 export { ButtonImage } from './components/buttons/button-image'
 export {
   ButtonQuestionMarkIcon,
@@ -130,6 +131,7 @@ export { ImageCropper } from './components/images/image-cropper'
 export { ImageWithCaption } from './components/images/image-with-caption'
 
 // Layout
+export { ArrowContainer } from './components/layout/arrow-container'
 export { HorizontalStroke } from './components/layout/horizontal-stroke'
 export { Line } from './components/layout/line'
 export { Marger } from './components/layout/marger'